Instantiation using Coordinators
Because:
- The 'Coordinator Pattern` will be the architecture used for this app
- Here I began to implement the coordinator pattern

This was Day 2 of 100DaysOfCode for building my portfolio. I started the
implementation of the coordinator pattern, it is far from complete. The app
does not work correctly at the moment. The master view has a nav controller
but this nav controller shows the detail view within it, instead of within
the detail view. I hope to resolve this tomorrow