Initial Commit of NTColorPickerViewController
Because:
- The NTColorPikcerViewController along with its view NTColorPickerView
  will be used to select colors
- Initially to verify the colors this is implemented as a color image,
  however this will be changed to a series of buttons

This is part of day 6 of 100DaysOfCode. The initial implementation was made
so as to allow the verification of colors. The implementation also shows
the colors in the same position in Dark Mode and Light Mode even though
the colors will change depending on whether the user is in light or dark
mode.