Improved NTColorPickerView
Because:
- The NTColorPickerView was improved so that it does not disappear until
  the user clicks outside the image square
- The NTColorPickerView now has a markerSquare to show the color currently
  selected
- The NTColorPickerView now sets its own constraints and dismisses itself
- The NTColorPickerDelegate was simplified so as to only have a color
  picked function

This was Day 8 of 100DaysOfCode. These changes almost complete the
NTColorPicker. What is missing now is to set the correct color for the
border and the marker square, show the marker square correctly when the
picker is first launched and calculate the color location automatically
based on the color selected. That is planned for Day 9