Added NTTextViewToolbar and necessary changes
Because:
- The NTTextViewToolbar appears above the keyboard and allows the user to
  dismiss they keyboard as well as set the text color and font.
- To complete its functionality it was necessary to modify slightly
  NTColorPickerView and its delegate
- Its implementation caused a number of changes in
  WrittenNotesViewController.

This was part of Day 11 of 100DaysOfCode. This was implemented to allow the
modification of the text. It is incomplete as you cannot set the text size
nor can you set the weight of the font. This may be added later but first I
want to implement more of the core functionality in terms of adding, saving
and deleting notes.