Added Note Model and Implementation
Because:
- This is a basic model of a note that will be used for testing and saved in
  core data
- Initially to simplify the work it is not using an attributed string
- The note is passed between the notelist view and the written notes view

The was day 13 of 100DaysOfCode. This is the first time a note is actually
being created. The implementation is incomplete as modifications are not
being correctly saved and there is no persistance. Also the name is not
appearing on the list when you return to the table view. This is necessary
in the final project.