NotesLixtView updated
Because:
- The notes list view is no longer a pure UITableViewController. This
  allows a button to easily be added to the view and means the navigation
  bar is no longer necessary
- The navigation bar is now hidden
- Extra cells at the bottom are hidden
- A rounded button was added to the bottom of the screen for adding notes.
- The cells now display the new title parameter instead of the text parameter

This was part of Day 25 of 100DaysOfCode. The navigation bar was on screen
just to show an add button, this has now been removed and a button has been
added to the view to act as the add new note button. The extra cells at the
bottom of the table view are also now hidden.