Added Remove Note feature and moved DataSource to seperate file
Because:
- The user must be able to delete unwanted notes.
- The data source was separated from the view controller as it was
  becomming quite long.

This was Day 27 of 100DaysOfCode. This feature completes the basic features
that the app needs in order to function. The user can now create a note,
modify it and delete it. As I do not have a developer account I cannot check
if the notes correctly sync across devices.